refactor(components/items): move `ItemsList` to `/list`
