+++ app/components/items/index.ts
@@ -1,10 +1,10 @@
 export * from './image'
 export * from './top-card'
-export * from './list'
+export * from './list/list'
 export * from './list/element'
 export * from './list/element.skeleton'
 export * from './image.skeleton'
-export * from './list.skeleton'
+export * from './list/list.skeleton'
 export * from './top-card.skeleton'
 export * from './relative-time'
 export * from './genre-chip'
